requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio[asyncio]==5.10.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

